[PIR #1205] Move REPLAY_BUFFER_MAX_BYTES to the buffer it configures
It was defined in shellper-protocol.ts beside REPLAY_PAYLOAD_MAX, on the
reasoning that they are the same wire-adjacent concern. They are not: a
retention ceiling is a memory policy, and shellper-protocol.ts declares
itself the wire-protocol module. It was filed there only because it is
derived from a wire constant.

Now lives next to the class it configures, still defined as
= REPLAY_PAYLOAD_MAX so the derivation stays self-maintaining if the send
cap moves. No new dependency: shellper-protocol.ts is under the same
built-ins-only constraint.
